requirements.txt
selenium>=4.15.0
requests>=2.31.0
beautifulsoup4>=4.12.2
streamlit>=1.28.0
openai>=1.40.1,<2.0.0
python-dotenv>=1.0.1,<2.0.0
pandas>=2.1.0
undetected-chromedriver>=3.5.4
